# [SPARK-6483][SQL]Improve ScalaUdf called performance.