--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/ScalaUdf.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/ScalaUdf.scala
@@ -703,4 +703,5 @@ case class ScalaUdf(function: AnyRef, dataType: DataType, children: Seq[Expressi
   // scalastyle:on
   
   override def eval(input: Row): Any = ScalaReflection.convertToCatalyst(f(input), dataType)
+
 }