@chenghao-intel 2015-03-24
Hmm, have you try what the performance gain by this change? From my understanding the bottleneck is in the function call `ScalaReflection.convertToScala`