@zzcclp 2015-03-24
Before this change, it takes 17 minutes, and now takes 5 minutes, which is the same as *HiveContext + udf floor* and *non-udf*