@chenghao-intel 2015-03-24
OK, probably we can also move the `children.size match {..}` out of the `eval`.