@zzcclp 2015-03-24
@chenghao-intel , I change code and test it, the result is the same as last commit  , is 5 minutes.
Please help me for checking code.