@chenghao-intel 2015-03-24
@zzcclp I will run the benchmark in my local machine, will get back soon.

@liancheng , can you trigger the unit test?