@chenghao-intel 2015-03-25
You need to fetch the latest code and resolve the conflicts.