@chenghao-intel 2015-03-25
@zzcclp LGTM, except some small issues.