@zzcclp 2015-03-25
@SparkQA , merge again.