@zzcclp 2015-03-25
@chenghao-intel , I have update the code. can you take a look again. thanks.

@AmplabJenkins , re-test