@chenghao-intel 2015-03-25
@zzcclp 
children:Seq[Expression] essentially can be either `List[Expression]` or `ArraySeq[Expression]`, the later doesn't support the pattern matching.

Probably you have to use the code like:
```scala
val f = children.size match {
  case 1 =>
   val func = function.asInstanceOf[(Any) => Any]
   val child0 = children(0)
    (input: Row) => {
      func(ScalaReflection.convertToScala(child0.eval(input), child0.dataType)))
    }

  case 2 =>
   val func = function.asInstanceOf[(Any) => Any]
   val child0 = children(0)
   val child1 = children(1)

   (input: Row) => {
    func(ScalaReflection.convertToScala(child0.eval(input), child0.dataType))
       ScalaReflection.convertToScala(child1.eval(input), child1.dataType)))
    }
}
```