@zzcclp 2015-03-25
@chenghao-intel , can you review again, thanks.