@zzcclp 2015-03-25
Add a newline at the end of source file, does it need to re test?