@zzcclp 2015-03-25
@AmplabJenkins , please re test, thanks.