@chenghao-intel 2015-03-25
The unit test will be auto-triggered, once the code changed, you needn't say anything to @AmplabJenkins .