@liancheng 2015-03-25
@zzcclp @chenghao-intel Thanks for working on this and the review comments! Merged to master.