@liancheng 2015-03-25
@zzcclp Would you please set your real name on both GitHub and JIRA so that our script can put your name on the credit list of the next release? Also, it would be good if you can set your name in git config:

```
$ git config --global user.name "Your Name"
```